[refactor] Dtos: split dto classes into separate files for sharing
