+++ composeApp/src/commonMain/kotlin/com/rwmobi/roctopus/data/source/network/ProductsEndpoint.kt
@@ -8,7 +8,7 @@
 package com.rwmobi.roctopus.data.source.network
 
 import com.rwmobi.roctopus.data.source.network.dto.ProductsApiResponse
-import com.rwmobi.roctopus.data.source.network.dto.StandardUnitRatesApiResponse
+import com.rwmobi.roctopus.data.source.network.dto.PricesApiResponse
 import com.rwmobi.roctopus.domain.extensions.formatInstantWithoutSeconds
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
@@ -53,7 +53,7 @@ class ProductsEndpoint(
         tariffCode: String,
         periodFrom: Instant? = null,
         periodTo: Instant? = null,
-    ): StandardUnitRatesApiResponse? {
+    ): PricesApiResponse? {
         return withContext(dispatcher) {
             httpClient.get("$endpointUrl/$productCode/electricity-tariffs/$tariffCode/standard-unit-rates") {
                 parameter("period_from", periodFrom?.formatInstantWithoutSeconds())
