+++ composeApp/src/commonMain/kotlin/com/rwmobi/roctopus/data/source/network/dto/LinkDto.kt
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) 2024. Ryan Wong
+ * https://github.com/ryanw-mobile
+ * Sponsored by RW MobiMedia UK Limited
+ *
+ */
+
+package com.rwmobi.roctopus.data.source.network.dto
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class LinkDto(
+    @SerialName("href") val href: String,
+    @SerialName("method") val method: String,
+    @SerialName("rel") val rel: String,
+)
